Clang code model related cleaning up (#1747).

+++ src/plugins/dataStore/BioSignalMLDataStore/src/biosignalmldatastoreexporter.cpp
@@ -95,7 +95,7 @@ void BiosignalmlDataStoreExporter::execute(QString &pErrorMessage) const
             bsml::HDF5::Clock::Ptr clock = recording->new_clock(recordingUri+"/clock/"+voi->uri().toStdString()+runNb,
                                                                 rdf::URI(baseUnits+voi->unit().toStdString()),
                                                                 voi->values(i),
-                                                                voi->size(i));
+                                                                size_t(voi->size(i)));
 
             clock->set_label(voi->label().toStdString());
 
@@ -136,7 +136,7 @@ void BiosignalmlDataStoreExporter::execute(QString &pErrorMessage) const
             double *dataPointer = data;
             int rowCount = 0;
 
-            for (quint64 j = 0, jMax = dataStore->size(i); j < jMax; ++j) {
+            for (int j = 0, jMax = dataStore->size(i); j < jMax; ++j) {
                 foreach (DataStore::DataStoreVariable *variable, variables)
                     *dataPointer++ = variable->value(j, i);
